Moved data and renamed OnlineSettingConfigurable
